pyproject.toml
[project]
name = "gl261-benchmark"
version = "0.1.0"
description = "Segmentation benchmark for the GL261 mouse brain tumor ultrasound dataset"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "segmentation-models-pytorch>=0.3",
    "scipy>=1.10",
    "numpy>=1.24",
    "pandas>=2.0",
    "Pillow>=9.0",
    "matplotlib>=3.7",
    "tqdm>=4.60",
    "requests>=2.28",
    "openpyxl>=3.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
include = ["gl261*"]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"
